API Active check added and a simple behaviour test.
I added a way to detect if the remote API is active.
The data is being passed to all application areas.
However the only part making use of it is the command at the moment. Obviously the most important part to check if the API is active or not.

Also a simple test added.